Add DTOs

+++ src/main/resources/db/migration/V001__cria-tabela-vendedores.sql
@@ -1,8 +1,8 @@
 CREATE TABLE vendedores
 (
-    id                   SERIAL      NOT NULL,
-    nome                 VARCHAR(60) NOT NULL,
-    total_vendas         BIGINT      NOT NULL,
-    media_vendas_diarias BIGINT      NOT NULL,
+    id                   SERIAL         NOT NULL,
+    nome                 VARCHAR(60)    NOT NULL,
+    total_vendas         BIGINT         NOT NULL,
+    media_vendas_diarias DECIMAL(10, 2) NOT NULL,
     PRIMARY KEY (id)
 );
